bug in group by fixed (let vars handling was not correct)

+++ src/compiler/parsetree/parsenodes.h
@@ -1503,7 +1503,7 @@ public:
 	void accept(parsenode_visitor&) const;
 
 public:
-  int size () const { return vardecl_hv.size (); }
+  size_t size () const { return vardecl_hv.size (); }
 };
